+++ osidb/__init__.py
@@ -2,5 +2,5 @@
 osidb version
 """
 
-__version__ = "4.15.0"
+__version__ = "4.16.0"
 VERSION = __version__