+++ docs/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add endpoint that checks if flaw is public or has its work completed (OSIDB-4328)
+
+### Changed
+- Deprecate `team` field from PsProduct (OSIDB-4319)
 
 ## [4.13.0] - 2025-07-14
 ### Added
@@ -15,7 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Filter affects based on absence/presence of related trackers (OSIDB-3893)
 - Allow excluding streams with existing trackers from tracker file offer (OSIDB-3893)
 - Add `cve_id` to affect and trackers endpoints (OSIDB-4293)
-- Add endpoint that checks if flaw is public or has its work completed (OSIDB-4328)
 
 ### Fixed
 - CVE.org and CISA CVSS correctly recalculate and persist the numeric