+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Emit a warning instead of an error when the PURL-derived PsComponent doesn't
+  match a user-provided PsComponent (OSIDB-4367)
 
 ## [4.14.0] - 2025-07-23
 ### Added