+++ docs/CHANGELOG.md
@@ -12,10 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fix FlawSource enum reference in OpenAPI schema (OSIDB-4401)
 - Fix the available-flaws API giving a 404 on non-existent flaws instead of 204 (OSIDB-4397).
+- Prevent emails from being sent for *rejected* critical flaws (OSIDB-4274)
 
 ### Changed
 - Emit a warning instead of an error when the PURL-derived PsComponent doesn't
   match a user-provided PsComponent (OSIDB-4367)
+- Initialize ACLs as internal (OSIDB-4274)
 
 ## [4.14.0] - 2025-07-23
 ### Added