+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fix `owner` update in Jira not reflected in OSIDB (OSIDB-4306)
+
 ### Changed
 - Make `affect` model's `flaw` field un-nullable (OSIDB-4202)