+++ docs/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [4.16.0] - 2025-09-16
 ### Fixed
 - Fix `owner` update in Jira not reflected in OSIDB (OSIDB-4306)
 - Fix not affected justifications from Jira not being collected by OSIDB (OSIDB-4418)